# docs/swagger.yaml
basePath: /
definitions:
  helpers.HTTPBadRequest:
    properties:
      error:
        properties:
          code:
            example: 400
            type: integer
          message:
            example: Field xxx wajib diisi.
            type: string
        type: object
    type: object
...
  helpers.HTTPForbidden:
    properties:
      error:
        properties:
          code:
            example: 403
            type: integer
          message:
            example: Pengguna tidak memiliki cukup izin untuk mengakses sumber daya.
            type: string
        type: object
    type: object
...
  helpers.HTTPUnauthorized:
    properties:
      error:
        properties:
          code:
            example: 401
            type: integer
          message:
            example: Token otentikasi tidak valid.
            type: string
        type: object
    type: object
...
  models.SettingShort:
    properties:
      db.migration.version:
        example: "0036"
        type: string
      db.seed.version:
        example: "0011"
        type: string
      origin_address.subdistrict.id:
        example: 222a761e-7dfc-4586-8f39-0d0fb59bb050
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: |-
    Zahir Courier Service API.

    ## Authorization

    This service use Zahir ID Oauth 2.0 Authorization, you SHOULD make authenticated requests with a zahir id access token using the `Authorization` request header field with the `Bearer` HTTP authorization scheme
    ```
    Authorization: Bearer {access_token}
    ```
    On this docs, click `Authorize` > fill the Value with `Bearer {access_token}` > click `Authorize`. Once you've done that, when you `Try it out`, the authorization header with your bearer token should be sent to the API.

    ## Query params

    By default, we support a common way for selecting fields, filtering, searching, sorting, and pagination in URL query params on `GET` method:

    ### Field

    Get selected fields in GET result, example:
    ```
    GET /api/resources?fields=field_a,field_b,field_c
    ```
    equivalent to sql:
    ```sql
    SELECT field_a, field_b, field_c FROM resources
    ```

    ### Filter

    Adds fields request condition (multiple conditions) to the request, example:
    ```
    GET /api/resources?field_a=value_a&field_b.$gte=value_b&field_c.$like=value_c&field_d.$ilike=value_d%
    ```
    equivalent to sql:
    ```sql
    SELECT * FROM resources WHERE (field_a = 'value_a') AND (field_b >= value_b) AND (field_c LIKE '%value_c%') AND (LOWER(field_d) LIKE LOWER('value_d%'))
    ```

    #### Available filter conditions

    * `$eq`: equal (`=`)
...

...

...

...
paths:
...
  /api/setting:
    get:
      consumes:
      - application/json
      description: Get current setting
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SettingShort'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.HTTPUnauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.HTTPForbidden'
      security:
      - ApiKeyAuth: []
      summary: Get current setting
      tags:
      - Setting
    post:
      consumes:
      - application/json
      description: Update current setting
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SettingShort'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SettingShort'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPBadRequest'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.HTTPUnauthorized'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/helpers.HTTPForbidden'
      security:
      - ApiKeyAuth: []
      summary: Update current setting
      tags:
      - Setting
...
schemes:
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"